# internal/trainings/ports/openapi_types.gen.go
// Package ports provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package ports

import (
	"time"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// PostTraining defines model for PostTraining.
type PostTraining struct {
	Notes string    `json:"notes"`
	Time  time.Time `json:"time"`
}

// Training defines model for Training.
type Training struct {
	CanBeCancelled     bool       `json:"canBeCancelled"`
	MoveProposedBy     *string    `json:"moveProposedBy,omitempty"`
	MoveRequiresAccept bool       `json:"moveRequiresAccept"`
	Notes              string     `json:"notes"`
	ProposedTime       *time.Time `json:"proposedTime,omitempty"`
	Time               time.Time  `json:"time"`
	User               string     `json:"user"`
	UserUuid           string     `json:"userUuid"`
	Uuid               string     `json:"uuid"`
}

// Trainings defines model for Trainings.
type Trainings struct {
	Trainings []Training `json:"trainings"`
}

// CreateTrainingJSONBody defines parameters for CreateTraining.
type CreateTrainingJSONBody PostTraining

// RequestRescheduleTrainingJSONBody defines parameters for RequestRescheduleTraining.
type RequestRescheduleTrainingJSONBody PostTraining

// RescheduleTrainingJSONBody defines parameters for RescheduleTraining.
type RescheduleTrainingJSONBody PostTraining

// CreateTrainingRequestBody defines body for CreateTraining for application/json ContentType.
type CreateTrainingJSONRequestBody CreateTrainingJSONBody

// RequestRescheduleTrainingRequestBody defines body for RequestRescheduleTraining for application/json ContentType.
type RequestRescheduleTrainingJSONRequestBody RequestRescheduleTrainingJSONBody

// RescheduleTrainingRequestBody defines body for RescheduleTraining for application/json ContentType.
type RescheduleTrainingJSONRequestBody RescheduleTrainingJSONBody